+++ packages/astro/src/@types/astro.ts
@@ -2,7 +2,7 @@ import type { AddressInfo } from 'net';
 import type * as babel from '@babel/core';
 import type * as vite from 'vite';
 import { z } from 'zod';
-import type { ShikiConfig, Plugin } from '@astrojs/markdown-remark';
+import type { ShikiConfig, RemarkPlugins, RehypePlugins } from '@astrojs/markdown-remark';
 import type { AstroConfigSchema } from '../core/config';
 import type { AstroComponentFactory, Metadata } from '../runtime/server';
 import type { ViteConfigWithSSR } from '../core/create-vite';
@@ -69,25 +69,110 @@ export interface BuildConfig {
 	server: URL;
 	serverEntry: string;
 	staticMode: boolean | undefined;
-	runtimeMarkdown: boolean;
 }
 
 /**
- * Astro.* available in all components
- * Docs: https://docs.astro.build/reference/api-reference/#astro-global
+ * Astro global available in all contexts in .astro files
+ *
+ * [Astro reference](https://docs.astro.build/reference/api-reference/#astro-global)
  */
 export interface AstroGlobal extends AstroGlobalPartial {
-	/** get the current canonical URL */
+	/** Canonical URL of the current page. If the [site](https://docs.astro.build/en/reference/configuration-reference/#site) config option is set, its origin will be the origin of this URL.
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astrocanonicalurl)
+	 */
 	canonicalURL: URL;
-	/** get page params (dynamic pages only) */
+	/** Parameters passed to a dynamic page generated using [getStaticPaths](https://docs.astro.build/en/reference/api-reference/#getstaticpaths)
+	 *
+	 * Example usage:
+	 * ```astro
+	 * ---
+	 * export async function getStaticPaths() {
+	 *    return [
+	 *     { params: { id: '1' } },
+	 *   ];
+	 * }
+	 *
+	 * const { id } = Astro.params;
+	 * ---
+	 * <h1>{id}</h1>
+	 * ```
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#params)
+	 */
 	params: Params;
-	/** set props for this astro component (along with default values) */
+	/** List of props passed to this component
+	 *
+	 * A common way to get specific props is through [destructuring](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Destructuring_assignment), ex:
+	 * ```typescript
+	 * const { name } = Astro.props
+	 * ```
+	 *
+	 * [Astro reference](https://docs.astro.build/en/core-concepts/astro-components/#component-props)
+	 */
 	props: Record<string, number | string | any>;
-	/** get information about this page */
+	/** Information about the current request. This is a standard [Request](https://developer.mozilla.org/en-US/docs/Web/API/Request) object
+	 *
+	 * For example, to get a URL object of the current URL, you can use:
+	 * ```typescript
+	 * const url = new URL(Astro.request.url);
+	 * ```
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astrorequest)
+	 */
 	request: Request;
-	/** see if slots are used */
+	/** Redirect to another page (**SSR Only**)
+	 *
+	 * Example usage:
+	 * ```typescript
+	 * if(!isLoggedIn) {
+	 *   return Astro.redirect('/login');
+	 * }
+	 * ```
+	 *
+	 * [Astro reference](https://docs.astro.build/en/guides/server-side-rendering/#astroredirect)
+	 */
+	redirect(path: string): Response;
+	/**
+	 * The <Astro.self /> element allows a component to reference itself recursively.
+	 *
+	 * [Astro reference](https://docs.astro.build/en/guides/server-side-rendering/#astroself)
+	 */
+	self: AstroComponentFactory;
+	/** Utility functions for modifying an Astro component’s slotted children
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astroslots)
+	 */
 	slots: Record<string, true | undefined> & {
+		/**
+		 * Check whether content for this slot name exists
+		 *
+		 * Example usage:
+		 * ```typescript
+		 *	if (Astro.slots.has('default')) {
+		 *   // Do something...
+		 *	}
+		 * ```
+		 *
+		 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astroslots)
+		 */
 		has(slotName: string): boolean;
+		/**
+		 * Asychronously renders this slot and returns HTML
+		 *
+		 * Example usage:
+		 * ```astro
+		 * ---
+		 * let html: string = '';
+		 * if (Astro.slots.has('default')) {
+		 *   html = await Astro.slots.render('default')
+		 * }
+		 * ---
+		 * <Fragment set:html={html} />
+		 * ```
+		 *
+		 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astroslots)
+		 */
 		render(slotName: string, args?: any[]): Promise<string>;
 	};
 }
@@ -96,12 +181,29 @@ export interface AstroGlobalPartial {
 	/**
 	 * @deprecated since version 0.24. See the {@link https://astro.build/deprecated/resolve upgrade guide} for more details.
 	 */
-	resolve: (path: string) => string;
-	/** @deprecated Use `Astro.glob()` instead. */
+	resolve(path: string): string;
+	/** @deprecated since version 0.26. Use [Astro.glob()](https://docs.astro.build/en/reference/api-reference/#astroglob) instead. */
 	fetchContent(globStr: string): Promise<any[]>;
+	/**
+	 * Fetch local files into your static site setup
+	 *
+	 * Example usage:
+	 * ```typescript
+	 * const posts = await Astro.glob('../pages/post/*.md');
+	 * ```
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astroglob)
+	 */
 	glob(globStr: `${any}.astro`): Promise<ComponentInstance[]>;
 	glob<T extends Record<string, any>>(globStr: `${any}.md`): Promise<MarkdownInstance<T>[]>;
 	glob<T extends Record<string, any>>(globStr: string): Promise<T[]>;
+	/**
+	 * Returns a [URL](https://developer.mozilla.org/en-US/docs/Web/API/URL) object built from the [site](https://docs.astro.build/en/reference/configuration-reference/#site) config option
+	 *
+	 * If `site` is undefined, the URL object will instead be built from `localhost`
+	 *
+	 * [Astro reference](https://docs.astro.build/en/reference/api-reference/#astrosite)
+	 */
 	site: URL;
 }
 
@@ -381,12 +483,22 @@ export interface AstroUserConfig {
 
 		/**
 		 * @docs
+		 * @name markdown.mode
+		 * @type {'md' | 'mdx'}
+		 * @default `mdx`
+		 * @description
+		 * Control wheater to allow components inside markdown files ('mdx') or not ('md').
+		 */
+		mode?: 'md' | 'mdx';
+
+		/**
+		 * @docs
 		 * @name markdown.shikiConfig
-		 * @type {ShikiConfig}
+		 * @typeraw {Partial<ShikiConfig>}
 		 * @description
 		 * Shiki configuration options. See [the markdown configuration docs](https://docs.astro.build/en/guides/markdown-content/#shiki-configuration) for usage.
 		 */
-		shikiConfig?: ShikiConfig;
+		shikiConfig?: Partial<ShikiConfig>;
 
 		/**
 		 * @docs
@@ -413,7 +525,7 @@ export interface AstroUserConfig {
 		/**
 		 * @docs
 		 * @name markdown.remarkPlugins
-		 * @type {Plugin[]}
+		 * @type {RemarkPlugins}
 		 * @description
 		 * Pass a custom [Remark](https://github.com/remarkjs/remark) plugin to customize how your Markdown is built.
 		 *
@@ -428,11 +540,11 @@ export interface AstroUserConfig {
 		 * };
 		 * ```
 		 */
-		remarkPlugins?: Plugin[];
+		remarkPlugins?: RemarkPlugins;
 		/**
 		 * @docs
 		 * @name markdown.rehypePlugins
-		 * @type {Plugin[]}
+		 * @type {RehypePlugins}
 		 * @description
 		 * Pass a custom [Rehype](https://github.com/remarkjs/remark-rehype) plugin to customize how your Markdown is built.
 		 *
@@ -447,7 +559,7 @@ export interface AstroUserConfig {
 		 * };
 		 * ```
 		 */
-		rehypePlugins?: Plugin[];
+		rehypePlugins?: RehypePlugins;
 	};
 
 	/**
@@ -516,13 +628,13 @@ export interface AstroUserConfig {
 
 	experimental?: {
 		/**
-		 * Enable experimental support for 3rd-party integrations.
+		 * Enable support for 3rd-party integrations.
 		 * Default: false
 		 */
 		integrations?: boolean;
 
 		/**
-		 * Enable a build for SSR support.
+		 * Enable support for 3rd-party SSR adapters.
 		 * Default: false
 		 */
 		ssr?: boolean;
@@ -654,12 +766,6 @@ export type JSXTransformFn = (options: {
 export interface ManifestData {
 	routes: RouteData[];
 }
-
-export type MarkdownRenderOptions = [string | MarkdownParser, Record<string, any>];
-export type MarkdownParser = (
-	contents: string,
-	options?: Record<string, any>
-) => MarkdownParserResponse | PromiseLike<MarkdownParserResponse>;
 
 export interface MarkdownParserResponse {
 	frontmatter: {