+++ packages/astro/src/runtime/server/index.ts
@@ -260,7 +260,7 @@ Did you mean to enable ${formatList(probableRendererNames.map((r) => '`' + r + '
 				// We already know that renderer.ssr.check() has failed
 				// but this will throw a much more descriptive error!
 				renderer = matchingRenderers[0];
-				({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children));
+				({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children, metadata));
 			} else {
 				throw new Error(`Unable to render ${metadata.displayName}!
 
@@ -279,7 +279,7 @@ If you're still stuck, please open an issue on GitHub or join us at https://astr
 		if (metadata.hydrate === 'only') {
 			html = await renderSlot(result, slots?.fallback);
 		} else {
-			({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children));
+			({ html } = await renderer.ssr.renderToStaticMarkup(Component, props, children, metadata));
 		}
 	}
 
@@ -290,7 +290,7 @@ If you're still stuck, please open an issue on GitHub or join us at https://astr
 			await render`<${Component}${spreadAttributes(props)}${markHTMLString(
 				(children == null || children == '') && voidElementNames.test(Component)
 					? `/>`
-					: `>${children}</${Component}>`
+					: `>${children == null ? '' : children}</${Component}>`
 			)}`
 		);
 	}
@@ -441,17 +441,32 @@ export function defineScriptVars(vars: Record<any, any>) {
 	return markHTMLString(output);
 }
 
-// Renders an endpoint request to completion, returning the body.
-export async function renderEndpoint(mod: EndpointHandler, request: Request, params: Params) {
-	const chosenMethod = request.method?.toLowerCase() ?? 'get';
-	const handler = mod[chosenMethod];
+function getHandlerFromModule(mod: EndpointHandler, method: string) {
+	// If there was an exact match on `method`, return that function.
+	if (mod[method]) {
+		return mod[method];
+	}
+	// Handle `del` instead of `delete`, since `delete` is a reserved word in JS.
+	if (method === 'delete' && mod['del']) {
+		return mod['del'];
+	}
+	// If a single `all` handler was used, return that function.
+	if (mod['all']) {
+		return mod['all'];
+	}
+	// Otherwise, no handler found.
+	return undefined;
+}
 
+/** Renders an endpoint request to completion, returning the body. */
+export async function renderEndpoint(mod: EndpointHandler, request: Request, params: Params) {
+	const chosenMethod = request.method?.toLowerCase();
+	const handler = getHandlerFromModule(mod, chosenMethod);
 	if (!handler || typeof handler !== 'function') {
 		throw new Error(
 			`Endpoint handler not found! Expected an exported function for "${chosenMethod}"`
 		);
 	}
-
 	return await handler.call(mod, params, request);
 }