+++ packages/renderers/renderer-lit/server-shim.js
@@ -1,5 +1,0 @@
-import { installWindowOnGlobal } from '@lit-labs/ssr/lib/dom-shim.js';
-installWindowOnGlobal();
-
-window.global = window;
-document.getElementsByTagName = () => [];