+++ packages/markdown/remark/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @astrojs/markdown-remark
 
+## 0.9.0
+
+### Minor Changes
+
+- [`53162534`](https://github.com/withastro/astro/commit/53162534450e160f65b95e7ef1523a106347ca28) Thanks [@FredKSchott](https://github.com/FredKSchott)! - - Removed `renderMarkdownWithFrontmatter` because it wasn't being used
+  - All options of `renderMarkdown` are now required — see the exported interface `AstroMarkdownOptions`
+  - New types: RemarkPlugin, RehypePlugin and ShikiConfig
+
+## 0.8.2
+
+### Patch Changes
+
+- [#2970](https://github.com/withastro/astro/pull/2970) [`b835e285`](https://github.com/withastro/astro/commit/b835e285defb4f31fc5ac1039c7f607c07f3c00b) Thanks [@JuanM04](https://github.com/JuanM04)! - Improved type checking
+
 ## 0.8.1
 
 ### Patch Changes